Take out UTC date conversion code
Need of UTC conversion was due to wrong configuration in the Test project, not necessary to actually implement. Date timezone will get saved depending on the project timezone.

If deemed necessary, it can be adjusted in the future to accommodate for user/browser variances in timezone (e.g. users that are not in EST) by calculating the timezone offset from local to EST and offsetting the displayed date by that.